+++ changelog.rst
@@ -1,11 +1,12 @@
 Changelog
 =========
 
-Last change: 18-AUG-2025 MTS
+Last change: 19-AUG-2025 MTS
 
 0.8.3
 -----
 - Removed ``lmfit`` dependency
+- Render by property allows different colormaps
 
 0.8.2
 -----